feat: Add monthly and annual subscription options to LandingPage
This commit adds the functionality to select between monthly and annual subscription options on the LandingPage. The user can now choose between the "Mensal" and "Anual" radio buttons, and the selected option is stored in the `planoAtivo` state variable. This allows for dynamic rendering of content based on the selected subscription plan.

Recent user commits:
- Update padding and font sizes in Header component
- chore: Update padding and font styles in Footer component
- chore: Update copyright year in Footer component
- chore: Update color variables in index.css file

Recent repository commits:
- markdown
- pomodoro
- Revert "Merge branch 'develop' into mongo"
- Merge branch 'mongo' of https://github.com/CosmicTasks/dev into mongo
- Merge branch 'develop' into mongo
- Merge branch 'mongo' of https://github.com/CosmicTasks/dev into mongo